move compliance tests into their own package
