+++ Cargo.toml
@@ -11,6 +11,7 @@ members = [
   "lib/sudo-pam-sys",
   "lib/sudo-pam",
   "lib/sudo-test",
+  "lib/sudo-compliance-tests",
 ]
 default-members = ["sudo"]
