+++ lib/sudo-test/Cargo.toml
@@ -8,6 +8,3 @@ doctest = false
 
 [dependencies]
 tempfile = "3.4.0"
-
-[dev-dependencies]
-pretty_assertions = "1.3.0"
